GeomShape {ggforce}	R Documentation

ggforce extensions to ggplot2

Description

ggforce makes heavy use of the ggproto class system to extend the functionality of ggplot2. In general the actual classes should be of little interest to users as the standard ggplot2 api of using geom_* and stat_* functions for building up the plot is encouraged.

[Package ggforce version 0.2.0 Index]