power_trans {ggforce}	R Documentation

Create a power transformation object

Description

This function can be used to create a proper trans object that encapsulates a power transformation (x^n).

Usage

power_trans(n)

Arguments

n
The degree of the power transformation

Value

A trans object

Examples

# Power of 5 transformations
trans <- power_trans(2)
trans$transform(1:10)

# Cubic root transformation
trans <- power_trans(1 / 3)
trans$transform(1:10)

# Use it in a plot
ggplot() +
  geom_line(aes(x = 1:10, y = 1:10)) +
  scale_x_continuous(trans = power_trans(2),
                     expand = c(0, 1))

[Package ggforce version 0.2.0 Index]